# perf(form-core): reduce instantiations when type checking