--- a/packages/form-core/src/metaHelper.ts
+++ b/packages/form-core/src/metaHelper.ts
@@ -106,7 +106,10 @@ export function metaHelper<
     const sortedFields = direction === 'up' ? fields : [...fields].reverse()
 
     sortedFields.forEach((fieldKey) => {
-      const nextFieldKey = updateIndex(fieldKey.toString(), direction)
+      const nextFieldKey = updateIndex(
+        (fieldKey as string).toString(),
+        direction,
+      )
       const nextFieldMeta = formApi.getFieldMeta(nextFieldKey)
       if (nextFieldMeta) {
         formApi.setFieldMeta(fieldKey, nextFieldMeta)